+++ AltaPayApi/AltaPayApi/Service/Requests/Common/PaymentRequestRequest.cs
@@ -19,6 +19,7 @@ namespace AltaPay.Service
 		public string OrganisationNumber { get; set; } // If the organisation_number parameter is given the organisation number field in the invoice payment form is prepopulated, and if no other payment options is enabled on the terminal the form will auto submit.
 		public AccountOffer AccountOffer { get; set; } // To require having account enabled for an invoice payment for this specific customer, set this to required. To disable account for this specific customer, set to disabled.
 		public PaymentSource Source { get; set; }
+		public AgreementConfig AgreementConfig { get; set; }